pyproject.toml
[project]
name = "mas"
version = "0.1.0"
description = "Multi-Agents Orchestration System: coordinate multiple coding CLIs as a role-based team"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12",
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "rich>=13.7",
]

[project.scripts]
mas = "mas.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mas"]

[tool.hatch.build.targets.wheel.force-include]
"templates" = "mas/_templates"

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
